netplan/requirements.txt
amqp==2.2.1
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-appconf==1.0.2
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-filter==1.0.4
django-libsass==0.7
django-material==1.0.0
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
djangorestframework==3.6.3
djangorestframework-csv==2.0.0
drf-proxy-pagination==0.1.1
html5lib==0.999999999
idna==2.5
kombu==4.1.0
libsass==0.13.2
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pip==9.0.1
psycopg2==2.7.3
pytz==2017.2
rcssmin==1.0.6
redis==2.10.5
requests==2.18.2
rjsmin==1.0.12
scipy==0.19.1
setuptools==36.2.4
six==1.10.0
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
vine==1.1.4
wagtail==1.11.1
webencodings==0.5.1
wheel==0.29.0
Willow==0.4
